requirements.txt
python-dotenv==1.0.0
slack-sdk==3.22.0